requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pdfplumber
sentence-transformers
faiss-cpu
tqdm
python-dotenv
google-generativeai
requests
